Moved interferometer unit tests into dedicated and expanded
